procedural sections and paragraphs
Instead of handling procedural sections and paragraphs headers as items
on par with the procedural statements, they are now handled like
containers, in the sense that procedural sections contain procedural
paragraphs, and these contain statements.

The first, and only the first, section of PROCEDURE DIVISION **CAN**
(though not need to) have no name.

The first, and only the first, paragraph of EACH section **CAN** (though
not need to) have no name.